+++ common/repo/inc/PinoutTable.h
@@ -1,7 +1,7 @@
 /*
  * ****************************************************************************
  *  @author      xqyjlj
- *  @file        pinout_table.h
+ *  @file        PinoutTable.h
  *  @brief
  *
  * ****************************************************************************
